+++ OAST/Server/IServer.cs
@@ -5,9 +5,9 @@ namespace OAST.Server
     public interface ICustomServer
     {
         public int QueueSize { get; set; }
-        int Mi { get; set; }
+        double Mi { get; set; }
         List<int> Queue { get; set; }
-        void SetMi(int mi);
+        void SetMi(double mi);
         void Reset();
         int Get();
         void SetQueueSize(int queueSize);
